Reduced extraneous files. All of the machine learning approaches are so short, they've all been put into the approach file. Also deleted old approach example, as it no longer applies to the current configuration. Began the live demo, which will perform live categorization as objects pass by, though there is still much to do.
